+++ lib/lacerda.rb
@@ -5,13 +5,14 @@ require 'lacerda/service'
 require 'lacerda/infrastructure'
 require 'lacerda/compare/json_schema'
 require 'lacerda/reporter'
+require 'lacerda/reporters/stdout'
 
 module Lacerda
   SCOPE_SEPARATOR = '::'
 
   def self.validate_reporter(reporter)
-    return true unless reporter
-    return true if reporter.class <= Lacerda::Reporter
+    return reporter unless reporter
+    return reporter if reporter.class <= Lacerda::Reporter
     raise "reporter must inherit from Lacerda::Reporter, but #{reporter.class.name} doesn't"
   end