+++ lib/lacerda/publish_contract.rb
@@ -11,7 +11,9 @@ module Lacerda
     def satisfies?(consumer, reporter = nil)
       Lacerda.validate_reporter(reporter)
       @comparator = Compare::JsonSchema.new(@schema)
-      @comparator.contains?(consumer.consume.scoped_schema(service), consumer.name)
+      result = @comparator.contains?(consumer.consume.scoped_schema(service), consumer.name)
+      reporter.try(:consume_specification_satisfied, consumer, result)
+      result
     end
 
     def object(name)