+++ CHANGELOG.markdown
@@ -1,5 +1,6 @@
 # 0.8.0 (30-Oct-15)
 - support for custom reporters
+- extract current output into and stdout reporter
 
 # 0.7.0 (30-Oct-15)
 - change ServiceName:ObjectName to ServiceName::ObjectName